requirements.txt
numpy
sounddevice
pyttsx3
python-dotenv
groq
faster-whisper
torch
tokenizers
requests
